Remove unnecessary dependence on ghc-8.8
This was convenient when GHC 8.8 was Homebrew's default version of GHC,
but now with M1 Mac CPUs in the mix it's a bit messy, and GHC has moved
on.

With this change, the EzPSL build files define what version of GHC to
use, and Stack downloads and uses the right one.  Build times are
longer and we can't use Homebrew's GHC if it's available, but such is
life.